+++ crates/ui/src/text/mod.rs
@@ -1,10 +1,12 @@
 mod format;
 mod inline;
 mod node;
+mod style;
 mod text_view;
 mod utils;
 
 use gpui::App;
+pub use style::*;
 pub use text_view::*;
 
 pub(crate) fn init(cx: &mut App) {
